Fix for ObjectExplorerUC - contextmenu focus issue
Context menu was getting focused, causing it to push a DispatcherFrame
and not release it.

Background:
The DispatcherFrame is pushed to support using the object explorer in
Excel which does not run a dispatcher. Menu items were not clickable if
below them was Excel (and not another WPF control). Pushing the
DispatcherFrame solves the problem.