Fixing ContextMenu+DispatcherFrame issue
When the context menu opens, we push a DispatcherFrame so we can
interact with the context menu in case it shows over Excel (in a VSTO
scenario). The context menu's Closing event does not fire if the
original item that owns the context menu is replaced by replacing the
DataContext or ItemsSource. To fix this, we call the contextMenu_Closing
handler manually when replacing the items, to ensure that the
DispatcherFrame is removed. Otherwise the lingering DispatcherFrame
causes all sorts of havoc (wait cursor while over Excel, stealing focus
so shortcuts in Excel don't work - Ctrl+T opens a new document in
QueryStorm instead of creating a new table in Excel, even though
an Excel cell is focused)