Removing BeginInvoke from TextBox_PreviewKeyDown
It was causing e.Handled to be ignored (since this was scheduled and not
executed) so that pressing enter when renaming a node in the object
explorer would activate the node in addition to renaming it.

Also, I see no reason why the dispatcher should be used when handling
key press at all, since this is guaranteed to be on the main UI thread.