Custom InitializeComponent() to use full assembly name
The problem this solves is with plugins that use a different version of
this assembly than the host. Both versions get loaded into the plugin's
appdomain so without specifying the full assembly name, the xaml can't
be loaded by the XamlReader.

Making sure that all custom dialogs and user controls defined in this
assembly use the full assembly name when loading resources ensures that
multiple versions of this assembly in the same appdomain don't cause
problems.